requirements.txt
click==4.0
Django==1.8.3
django-cors-headers==1.1.0
django-filter==0.10.0
django-solo==1.1.0
djangorestframework==3.1.3
jsonfield==1.0.3
Markdown==2.6.2
python-dotenv==0.1.2
wheel==0.24.0
